Move M3 theming stuff to custom-preferences-theming-integration
The M3 styles in this library only included the layout and not the text styling that makes them look actually like M3. The styling instead is provided in theming-preference-integration. This is surprising behavior for people only using the core lib and not the custom-preferences-theming-integration lib. I would rather remove M3 theming capability from the lib than leave it in a half-baked state.